+++ docs/whatsnew.rst
@@ -11,6 +11,7 @@ The full release history can be viewed `at the GitHub yank releases page <https:
 - Online analysis will always run by default now, with no target error, run every checkpoint interval, and with at least 200 iterations
 - Online analysis can now be a set to the checkpoint interval by setting ``online_analysis_interval = "checkpoint"`` in the ``MultiStateSampler``
 - Checkpoint interval increased from default of 10 to 200
+- Analysis now uses the online-analysis data if available by default
 
 0.22.0 RMSD the Casbah
 ----------------------